imtihon_7_month/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
idna==3.6
Pillow==10.1.0
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
telebot==0.0.5
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wikipedia==1.4.0
Wikipedia-API==0.6.0
